+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [9.8.1] 2020-11-05
+## Modified
+- PriceInputComponent : changed property name from selectValue to optional param selectOnFocus 
+- Modified PriceInput Component to handle selection of input value on focus
+
 ## [9.8.0] 2020-11-05
 ## Added
 - Added outlined material icon font to styles